+++ laterpay/compat.py
@@ -7,6 +7,7 @@
 from __future__ import absolute_import, print_function
 
 import sys
+import six
 
 py3k = sys.version_info[:2] >= (3, 0)
 
@@ -61,3 +62,14 @@ else:
 
     from urllib2 import Request, urlopen, URLError
 
+
+def encode_if_unicode(value, encoding='utf-8'):
+    """
+    Encode and return a ``value`` using specified ``encoding``.
+
+    Encoding is done only if ``value`` is a ``unicode`` instance
+    (utf-8 encoding is used as default).
+    """
+    if six.PY2 and isinstance(value, unicode):
+        value = value.encode(encoding)
+    return value